# frontend/src/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "LifeGrid API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/sites": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "List Sites",
    "operationId": "list_sites_api_v1_sites_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SiteOut"
         },
         "type": "array",
         "title": "Response List Sites Api V1 Sites Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "sites"
    ],
    "summary": "Create Site",
    "operationId": "create_site_api_v1_sites_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SiteOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}": {
   "patch": {
    "tags": [
     "sites"
    ],
    "summary": "Update Site",
    "operationId": "update_site_api_v1_sites__site_id__patch",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Site Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SitePatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SiteOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "SiteIn": {
    "properties": {
     "code": {
      "type": "string",
      "pattern": "^[A-Z0-9-]{2,16}$",
      "title": "Code"
     },
     "name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 2,
      "title": "Name"
     },
     "type": {
      "$ref": "#/components/schemas/SiteType"
     },
     "lat": {
      "type": "number",
      "maximum": 90.0,
      "minimum": -90.0,
      "title": "Lat"
     },
     "lng": {
      "type": "number",
      "maximum": 180.0,
      "minimum": -180.0,
      "title": "Lng"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone",
      "default": "Asia/Dhaka"
     }
    },
    "type": "object",
    "required": [
     "code",
     "name",
     "type",
     "lat",
     "lng"
    ],
    "title": "SiteIn"
   },
   "SiteOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "type": {
      "$ref": "#/components/schemas/SiteType"
     },
     "lat": {
      "type": "number",
      "title": "Lat"
     },
     "lng": {
      "type": "number",
      "title": "Lng"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone"
     },
     "active": {
      "type": "boolean",
      "title": "Active"
     }
    },
    "type": "object",
    "required": [
     "id",
     "code",
     "name",
     "type",
     "lat",
     "lng",
     "timezone",
     "active"
    ],
    "title": "SiteOut"
   },
   "SitePatch": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120,
        "minLength": 2
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "lat": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 90.0,
        "minimum": -90.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Lat"
     },
     "lng": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 180.0,
        "minimum": -180.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Lng"
     },
     "timezone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Timezone"
     },
     "active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Active"
     }
    },
    "type": "object",
    "title": "SitePatch"
   },
   "SiteType": {
    "type": "string",
    "enum": [
     "hospital",
     "blood_bank"
    ],
    "title": "SiteType"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}